requirements.txt
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
PyYAML==6.0.1
kaggle==1.5.16
awscli==1.29.67
psutil==5.9.6 